+++ cmd/add_tool.go
@@ -57,7 +57,7 @@ func runAddTool(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	config, err := generator.GetCurrentProjectConfig()
+	config, err := generator.GetCurrentProjectConfig(workingDir)
 	if err != nil {
 		return fmt.Errorf("failed to detect project: %w\nMake sure you're in a hexagonal architecture project directory", err)
 	}