+++ pkg/utils/parse.go
@@ -1,0 +1,11 @@
+package utils
+
+import "strconv"
+
+func ParseInt(s string) int {
+	i, e := strconv.Atoi(s)
+	if e != nil {
+		return 0
+	}
+	return i
+}